refactor: split 6 large files into focused sub-modules
- rules-command-operations.ts: 727→58 (barrel; types, CRUD, config, migration ops)
- ai.ts: 763→18 (barrel; completion, model-cache, provider command groups)
- config.ts: 677→30 (barrel; helpers, list/show, get/set, validate/doctor)
- deps-rendering.ts: 637→14 (barrel; ASCII rendering vs Graphviz rendering)
- formatting-utilities.ts: 636→23 (barrel; config formatting split by concern)
- taskCommands.ts: 694→31 (barrel; query vs mutation command groups)

All exports preserved via re-export barrels. No behavioral changes.
